+++ playground/client.js
@@ -2,8 +2,4 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Component from './index';
 
-ReactDOM.render(
-  React.createElement(
-    Component
-  ), document
-);
+ReactDOM.render(React.createElement(Component), document);
