+++ playground/server.js
@@ -1,18 +1,20 @@
 import React from 'react';
-import Playground from './index.js';
 import express from 'express';
 import ReactDOMServer from 'react-dom/server';
+
+import Playground from './Playground';
+
 
 const app = express();
 const port = 8000;
 
-app.get('/', function(req, resp) {
+app.get('/', (req, resp) => {
   const html = ReactDOMServer.renderToString(
-    React.createElement(Playground)
+    React.createElement(Playground),
   );
   resp.send(html);
 });
 
-app.listen(port, function() {
+app.listen(port, () => {
   console.log(`http://localhost:${port}`); // eslint-disable-line no-console
 });
